issue #74
