+++ code_templates/get_keys.py
@@ -1,0 +1,4 @@
+# get the labels of the data columns
+def get_code(args):
+   keys = args[0].keys()
+   return keys
