+++ code_templates/pandas/clean_data.py
@@ -1,0 +1,4 @@
+# removes rows with null entries
+def get_code(args):
+   args[0].dropna(axis=0,inplace=True)
+   return args[0]
