+++ code_templates/pandas/describe_data.py
@@ -1,0 +1,4 @@
+# describes a summary of a given dataframe
+def get_code(args):
+   data_description = args[0].describe()
+   return data_description
