+++ code_templates/pandas/drop_x.py
