+++ code_templates/pandas/select_y.py
@@ -1,0 +1,5 @@
+# split the data set into training and test data
+def get_code(args):
+   y_values = args[0][args[1]]
+   x_values = args[0].drop(args[1], axis = 1)
+   return (x_values, y_values)
