+++ code_templates/sklearn/eval_lin_reg.py
