+++ code_templates/sklearn/train_decision_tree.py
