+++ code_templates/sklearn/train_lin_reg.py
