+++ code_templates/tensorflow/predict_lin_reg.py
