--- pandas_code/code_templates/clean_data.py
@@ -1,4 +1,0 @@
-# removes rows with null entries
-def get_code(args):
-   args[0].dropna(axis=0,inplace=True)
-   return args[0]
