--- pandas_code/code_templates/describe_data.py
@@ -1,4 +1,0 @@
-# describes a summary of a given dataframe
-def get_code(args):
-   data_description = args[0].describe()
-   return data_description
