--- pandas_code/code_templates/get_keys.py
@@ -1,4 +1,0 @@
-# get the labels of the data columns
-def get_code(args):
-   keys = args[0].keys()
-   return keys
