--- pandas_code/code_templates/read_csv.py
@@ -1,6 +1,0 @@
-#  file into a dataframe for later processing
-import pandas as pd
-def get_code(args):
-   dataframe = pd.read_csv(args[0])
-   print(dataframe.dtypes)
-   return dataframe
