+++ pandas_code/parse_template.py
@@ -1,8 +1,17 @@
 import re
 import os
 def parse_template(template_name, args):
+   # hashmap for access of PATH
+   root_map = {'clean_data' : 'pandas/', 'describe_data' : 'pandas/',
+               'drop_x' : 'pandas/', 'one_hot_encode_x_data' : 'pandas/',
+               'one_hot_encode_y_data' : 'pandas/', 'read_csv' : 'pandas/', 
+               'select_y' : 'pandas/', 'eval_decision_tree' : 'sklearn/', 
+               'eval_lin_reg' : 'sklearn/','split' : 'sklearn/',
+               'train_decision_tree' : 'sklearn/', 'train_lin_reg' : 'sklearn/',
+               'get_keys' : '','predict_decision_tree' : 'tensorflow/', 'predict_lin_reg' : 'tensorflow/' }
+   
    root_dir = os.environ.get("PYTHONPATH", None)
-   template = root_dir+'/pandas_code/code_templates/'+template_name+".py"
+   template = root_dir+'/code_templates/'+root_map[template_name]+template_name+".py"
    generated_comments = []
    generated_code = []
    indent_size = 0
